requirements.txt
flask
msal
azure-storage-blob
